Make strtosz() return int64_t instead of ssize_t
RH-Author: Jes Sorensen <Jes.Sorensen@redhat.com>
Message-id: <1296476610-28514-22-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 17316
O-Subject: [PATCH 21/26] Make strtosz() return int64_t instead of ssize_t
Bugzilla: 637701
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>

From: Jes Sorensen <Jes.Sorensen@redhat.com>

strtosz() needs to return a 64 bit type even on 32 bit
architectures. Otherwise qemu-img will fail to create disk
images >= 2GB

(cherry picked from commit 70b4f4bb05ff5e6812c6593eeefbd19bd61b517d)
---
 cutils.c      |    8 ++++----
 qemu-common.h |    4 ++--
 qemu-img.c    |    2 +-
 vl.c          |    4 ++--
 4 files changed, 9 insertions(+), 9 deletions(-)

+++ qemu-common.h
@@ -142,8 +142,8 @@ int qemu_fdatasync(int fd);
 #define STRTOSZ_DEFSUFFIX_MB	'M'
 #define STRTOSZ_DEFSUFFIX_KB	'K'
 #define STRTOSZ_DEFSUFFIX_B	'B'
-ssize_t strtosz(const char *nptr, char **end);
-ssize_t strtosz_suffix(const char *nptr, char **end, const char default_suffix);
+int64_t strtosz(const char *nptr, char **end);
+int64_t strtosz_suffix(const char *nptr, char **end, const char default_suffix);
 
 /* path.c */
 void init_paths(const char *prefix);
